requirements.txt
cachetools~=5.5.2
certifi~=2025.6.15
charset-normalizer~=3.4.2
DAWG2-Python~=0.9.0
google-auth~=2.40.3
google-auth-oauthlib~=1.2.2
gspread~=6.2.1
httplib2~=0.22.0
idna~=3.10
oauth2client~=4.1.3
oauthlib~=3.2.2
pillow~=11.1.0
pyasn1~=0.6.1
pyasn1_modules~=0.4.2
pymorphy3~=2.0.3
pymorphy3-dicts-ru~=2.4.417150.4580142
pyparsing~=3.2.3
pyTelegramBotAPI~=4.27.0
requests~=2.32.4
requests-oauthlib~=2.0.0
rsa~=4.9.1
setuptools~=75.8.2
six~=1.17.0
telebot~=0.0.5
urllib3~=2.4.0
